+++ lib/to_currency_select_tag.rb
@@ -11,9 +11,9 @@ module ActionView
         html_options = html_options.stringify_keys
         add_default_name_and_id(html_options)
         value = if method(:value).arity.zero?
-                  value()
+                  options.fetch(:selected) { value() }
                 else
-                  value(object)
+                  options.fetch(:selected) { value(object) }
                 end
         content_tag(
           'select',