+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z (unreleased)
 
 ### Added
+- Add official support for Ruby 3.2 (#125)
 
 ### Changed
 - Drop whitespace between HTML tags (#115)