+++ CHANGELOG.md
@@ -15,7 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Removed
 - Drop support for Rails 5.2 (#112)
-- Drop support for Ruby 2.6 (#111)
+- Drop support for Ruby 2.6 and 2.7 (#111, #132)
 
 ### Fixed